requirements.txt
fastapi>=0.104.0
sqlalchemy>=2.0.0
pydantic>=2.5.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt==4.2.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
psycopg2-binary>=2.9.0
mangum>=0.19.0